project-root/web_master/requirements.txt
flask==3.0.0
requests==2.31.0
pyyaml==6.0.1